requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=3.13.0